+++ packages/bcgov-theme/stories/card/args.ts
@@ -1,0 +1,5 @@
+export const argTypes = {
+  title: {
+    description: 'The title of the card to display in its header.',
+  },
+};
